[WIP][JIT] Add `torch.jit.dict`
**Summary**
This commit adds `torch.jit.dict`, a dictionary type that has reference
semantics across the Python/TorchScript boundary. That is, modifications
made to instances of `torch.jit.dict` in TorchScript are visible in
Python even when it is not returned from the function.

`torch.jit.dict` is implemented using a modified version of pybind's `stl_bind.h`-style bindings attached to `ScriptDict`, `ScriptDictIterator` and `ScriptDictKeyIterator`, wrapper classes around `c10::impl::GenericDict` and `c10::impl::GenericDict::iterator`. These bindings allow instances of `torch.jit.dict` to be used as if it were a regular `dict` Python. Reference semantics are achieved by simply retrieving the `IValue` contained in `ScriptDict` in `toIValue` (invoked when converting Python arguments to `IValues` before calling TorchScript code).

**Test Plan**
This commit adds `TestScriptDict` to `test_list_dict.py`, a set of tests
that check that all of the common dictionary operations are supported
and that instances have reference semantics across the
Python/TorchScript boundary.

ghstack-source-id: 69064501367d42ba40b5553f909f8b94563cc6dd
Pull Request resolved: https://github.com/pytorch/pytorch/pull/52659